[IMPROV] Cleanup value objects

+++ src/test/java/ca/ulaval/glo2003/beds/domain/helpers/PackageObjectMother.java
@@ -2,10 +2,9 @@ package ca.ulaval.glo2003.beds.domain.helpers;
 
 import static ca.ulaval.glo2003.interfaces.helpers.Randomizer.randomEnum;
 
-import ca.ulaval.glo2003.beds.domain.*;
+import ca.ulaval.glo2003.beds.domain.Packages;
 import ca.ulaval.glo2003.transactions.domain.Price;
 import com.github.javafaker.Faker;
-import java.math.BigDecimal;
 
 public class PackageObjectMother {
 
@@ -17,6 +16,6 @@ public class PackageObjectMother {
 
   public static Price createPricePerNight() {
     double randomDouble = Faker.instance().number().randomDouble(2, 100, 1000);
-    return new Price(BigDecimal.valueOf(randomDouble));
+    return new Price(randomDouble);
   }
 }
