+++ src/test/java/ca/ulaval/glo2003/bookings/domain/CancelationRefundCalculatorTest.java
@@ -3,18 +3,17 @@ package ca.ulaval.glo2003.bookings.domain;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import ca.ulaval.glo2003.transactions.domain.Price;
-import java.math.BigDecimal;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 public class CancelationRefundCalculatorTest {
 
   private static CancelationRefundCalculator cancelationRefundCalculator;
-  private static Price evenTotal = new Price(BigDecimal.valueOf(20.00));
-  private static Price halfEvenTotal = new Price(BigDecimal.valueOf(10.00));
-  private static Price oddTotal = new Price(BigDecimal.valueOf(44.11));
-  private static Price higherHalfOddTotal = new Price(BigDecimal.valueOf(22.06));
-  private static Price lowerHalfOddTotal = new Price(BigDecimal.valueOf(22.05));
+  private static Price evenTotal = new Price(20.00);
+  private static Price halfEvenTotal = new Price(10.00);
+  private static Price oddTotal = new Price(44.11);
+  private static Price higherHalfOddTotal = new Price(22.06);
+  private static Price lowerHalfOddTotal = new Price(22.05);
 
   @BeforeAll
   public static void setUpCalculator() {
